Added the integration tests for TodoController

+++ src/main/java/com/example/to_do_list/controller/TodoController.java
@@ -45,6 +45,8 @@ public class TodoController {
 
     @DeleteMapping("/task/{id}")
     public ResponseEntity<Void> deleteTask(@PathVariable Long id) {
+        if (taskService.getById(id).isEmpty())
+            return ResponseEntity.notFound().build();
         taskService.delete(id);
         return ResponseEntity.noContent().build();
     }
